Implements #320 16 bit normals

+++ components/operators/include/ftl/operators/mvmls.hpp
@@ -18,7 +18,7 @@ class MultiViewMLS : public ftl::operators::Operator {
 	private:
 	std::vector<ftl::cuda::TextureObject<float4>*> centroid_horiz_;
 	std::vector<ftl::cuda::TextureObject<float4>*> centroid_vert_;
-	std::vector<ftl::cuda::TextureObject<float4>*> normals_horiz_;
+	std::vector<ftl::cuda::TextureObject<half4>*> normals_horiz_;
     std::vector<ftl::cuda::TextureObject<float>*> contributions_;
 };
